# api/openapi.yaml
openapi: "3.0.3"

info:
  title: Auth Service API
  version: "1.0.0"
  description: Authentication and session management service

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
  - name: Auth

paths:
  /v1/health:
    get:
      summary: Health check
      operationId: health
      tags: [Health]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/auth/signup:
    post:
      summary: Sign up a new user
      operationId: signup
      tags: [Auth]
      description: |
        Registers a new user for a given app.

        **Validations (in order):**
        1. All fields must be present and non-empty.
        2. `app_id` must be a valid UUID and the app must exist in the database.
        3. `email` must be a valid email address (RFC 5322).
        4. `password` must satisfy all password rules.
        5. The email must not already be registered for the given app.

        **Password Rules:**
        - 8–128 characters
        - At least one uppercase letter, one lowercase letter, one digit, and one special character
        - No leading or trailing spaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
            example:
              app_id: "00000000-0000-0000-0000-000000000001"
              email: "user@example.com"
              password: "StrongPass1!"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  value:
                    error: "app_id, email, and password are required"
                invalid_app_id:
                  value:
                    error: "invalid app_id"
                invalid_email:
                  value:
                    error: "invalid email address"
                password_length:
                  value:
                    error: "password must be 8-128 characters"
                password_complexity:
                  value:
                    error: "password must contain at least one uppercase letter, one lowercase letter, one number, and one special character"
                password_spaces:
                  value:
                    error: "password must not have leading or trailing spaces"
        "404":
          description: App not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "app not found"
        "409":
          description: Email already registered for this app
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "email already registered for this app"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/login:
    post:
      summary: Authenticate user and create session
      operationId: login
      tags: [Auth]
      description: |
        Authenticates a user and returns access and refresh tokens.

        **Validations (in order):**
        1. All fields must be present and non-empty.
        2. `app_id` must be a valid UUID and the app must exist in the database.
        3. `email` must be a valid email address (RFC 5322).

        **Business Logic:**
        - Looks up the user by `app_id` + `email`.
        - Checks the user is active.
        - Verifies password against stored bcrypt hash.
        - Creates a session (24h expiry) with User-Agent and IP.
        - Generates a refresh token (30-day expiry, SHA-256 hash stored).
        - Returns a JWT access token (RS256, 15-minute expiry).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              app_id: "00000000-0000-0000-0000-000000000001"
              email: "user@example.com"
              password: "StrongPass1!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  value:
                    error: "app_id, email, and password are required"
                invalid_app_id:
                  value:
                    error: "invalid app_id"
                invalid_email:
                  value:
                    error: "invalid email address"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid email or password"
        "404":
          description: App not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "app not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/token/refresh:
    post:
      summary: Rotate refresh token and issue new access token
      operationId: refreshToken
      tags: [Auth]
      description: |
        Rotates the refresh token and issues a new access token.

        **Business Logic:**
        1. Finds the refresh token by its SHA-256 hash (must not be revoked or expired).
        2. Verifies the associated session is not revoked and not expired.
        3. Revokes the old refresh token.
        4. Creates a new refresh token linked to the same session and user.
        5. Generates a new JWT access token (RS256, 15-minute expiry).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refresh_token: "hex-encoded-refresh-token"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "400":
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "refresh_token is required"
        "401":
          description: Token invalid, revoked, or session expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                token_not_found:
                  value:
                    error: "refresh token not found or expired"
                session_revoked:
                  value:
                    error: "session has been revoked"
                session_expired:
                  value:
                    error: "session has expired"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/logout:
    post:
      summary: Logout from current session
      operationId: logout
      tags: [Auth]
      description: |
        Revokes the current session and its associated refresh tokens.
        The session ID is extracted from the JWT access token by the auth middleware.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "logged out successfully"
        "400":
          description: Invalid session ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid session_id"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid or expired token"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/logout-all:
    post:
      summary: Logout from all devices
      operationId: logoutAll
      tags: [Auth]
      description: |
        Logs the user out from all devices by revoking all sessions and refresh tokens.
        The user ID is extracted from the JWT access token by the auth middleware.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out from all devices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "logged out from all devices"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid or expired token"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/password/change:
    post:
      summary: Change password
      operationId: changePassword
      tags: [Auth]
      description: |
        Changes the password for the authenticated user. Revokes all existing sessions
        and refresh tokens after the change.

        **Validations:**
        1. Both `current_password` and `new_password` must be present and non-empty.
        2. `current_password` must match the stored bcrypt hash.
        3. `new_password` must satisfy all password rules (same as signup).

        **Business Logic:**
        1. Fetches the current password hash for the authenticated user.
        2. Verifies `current_password` against the stored bcrypt hash.
        3. Validates `new_password` against password rules.
        4. Hashes the new password with bcrypt and updates the user record.
        5. Revokes all sessions for the user.
        6. Revokes all refresh tokens for the user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
            example:
              current_password: "OldStrongPass1!"
              new_password: "NewStrongPass1!"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "password changed successfully"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  value:
                    error: "current_password and new_password are required"
                password_length:
                  value:
                    error: "password must be 8-128 characters"
                password_complexity:
                  value:
                    error: "password must contain at least one uppercase letter, one lowercase letter, one number, and one special character"
                password_spaces:
                  value:
                    error: "password must not have leading or trailing spaces"
        "401":
          description: Missing/invalid token or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_token:
                  value:
                    error: "invalid or expired token"
                wrong_password:
                  value:
                    error: "invalid email or password"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "user not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok

    SignupRequest:
      type: object
      required: [app_id, email, password]
      properties:
        app_id:
          type: string
          format: uuid
          description: UUID of the app (must exist in database)
        email:
          type: string
          format: email
          description: Valid email address (RFC 5322)
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: |
            Must contain at least one uppercase letter, one lowercase letter,
            one digit, and one special character. No leading or trailing spaces.

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        app_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_active:
          type: boolean
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [app_id, email, password]
      properties:
        app_id:
          type: string
          format: uuid
          description: UUID of the app (must exist in database)
        email:
          type: string
          format: email
          description: Valid email address (RFC 5322)
        password:
          type: string
          minLength: 1

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: RS256 JWT access token (15-minute expiry)
        refresh_token:
          type: string
          description: Opaque refresh token (30-day expiry)

    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Current valid refresh token

    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: New RS256 JWT access token (15-minute expiry)
        refresh_token:
          type: string
          description: New opaque refresh token (30-day expiry)
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 900
        token_type:
          type: string
          description: Token type
          example: Bearer

    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
          description: The user's current password
        new_password:
          type: string
          minLength: 8
          maxLength: 128
          description: |
            Must contain at least one uppercase letter, one lowercase letter,
            one digit, and one special character. No leading or trailing spaces.

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string